'setuptools' (installable Python module) restructuring; extracted DataExtractor_* classes to separate files
