+++ README.md
@@ -44,15 +44,15 @@ Running `grunt build` will build the site based on the data currently in the `bu
 
 #### Update data then build site
 
-Running `grunt update` will fetch JSON for resource records, resource record trees and archival objects from ArchivesSpace
-using `utilities/getJson.py` and save it in your `build/data` directory, then will build the site based on that data.
+Running `grunt update` will fetch JSON for resource records, resource record trees and archival objects from ArchivesSpace 
+using `static_aid/getJson.py` and save it in your `build/data` directory, then will build the site based on that data.
 
 **WARNING**: Depending on the size of your ArchivesSpace installation, it could take quite a while for this script to
 loop through all resource records and components. Be patient!
 
 #### Clean Build
 
-By default, `grunt update` will only fetch JSON updated since the last time `utilities/getJson.py` completed successfully.
+By default, `grunt update` will only fetch JSON updated since the last time `static_aid/getJson.py` completed successfully. 
 At any point, you can run `grunt rebuild` to wipe out the existing data and build the site from scratch.
 
 **WARNING**: Depending on the size of your ArchivesSpace installation, it could take quite a while for this script to
