+++ static_aid/__init__.py
@@ -1,5 +1,1 @@
 # StaticAid Python Module
-
-# NOTE: not using __init__ features because this code base does not conform to Python standards
-# (there is no "root" from which package names are determined). Instead, individual modules
-# can be imported by name (see "import config" lines in *.py).
